[Port] Bunch of QoL, new stuff and fixes from Azure Peak (#1844)
1. Fixed bug metal heat with tongs.

2. Possibility to add material from tongs to recipe on anvil.
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/381

3. Added confirm for tabards, shields. "You get a prompt to confirm
whether you want to keep your shield heraldry. You also have to right
click shields to set their heraldry now." - EvenInDeathIStillServe
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/100

4. Added Preload Ambience feature
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/331

5. Added bold text for weapons when examining characters
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/75

6. Added wall clock functionality (show the time and day)
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/101

7. Ported one-click collection of arrows and bolts into quiver.
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/327

8. Added sounds of weapon draw and stow
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/349

9. Added the ability to drink from a fountain
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/350

10. Increase fuel for greater furnace. "You can fill up smelters with
fuel up to 30 minutes. One coal is 10 minutes." - EvenInDeathIStillServe
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/370

11. Fix armour damage (upper armour takes damage, then lower armour)
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/326

12. Ported autosmithing, it's less miserable for smiths players.
"Striking an ingot will begin to automatically continue striking, until
you either fail a strike, the bar cools, or the item finishes. You can
still manually strike. Auto smithing take 4 times as many strikes as
manual smithing." - EvenInDeathIStillServe
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/294

Feat. SayonaraEcho


![image](https://github.com/user-attachments/assets/b7fc7309-466e-4120-ab2a-c24c446ccaea)

![image](https://github.com/user-attachments/assets/e24bc279-552c-486c-b4fa-d00827442f9c)

![image](https://github.com/user-attachments/assets/74efef7f-b75c-459a-bc0b-ceefb1d3896c)

![image](https://github.com/user-attachments/assets/15fee833-44a6-4c5e-967b-4bca60cc2731)

+++ code/game/objects/lighting/roguelighting.dm
@@ -91,12 +91,16 @@
 	var/cookonme = FALSE
 	var/crossfire = TRUE
 	var/can_damage = FALSE
+	var/start_fuel //Override for fueluse. Mostly used for smelters.
+	var/fuel_modifier = 1 //Modifier for firefuel
 
 /obj/machinery/light/rogue/Initialize()
 	if(soundloop)
 		soundloop = new soundloop(list(src), FALSE)
 		soundloop.start()
 	GLOB.fires_list += src
+	if(start_fuel)
+		fueluse = start_fuel
 	if(fueluse)
 		fueluse = fueluse - (rand(fueluse*0.1,fueluse*0.3))
 	update_icon()
@@ -228,19 +232,21 @@
 		if(initial(fueluse))
 			if(fueluse > initial(fueluse) - 5 SECONDS)
 				to_chat(user, span_warning("The fire is fully fueled."))
-				return
+				return FALSE
 		else
 			if(!on)
 				return
 		if (alert(usr, "Feed [W] to the fire?", "ROGUETOWN", "Yes", "No") != "Yes")
 			return
+		if(!W)
+			return
 		qdel(W)
 		user.visible_message(span_warning("[user] feeds [W] to [src]."))
 		if(initial(fueluse))
-			fueluse = fueluse + W.firefuel
+			fueluse = fueluse + W.firefuel*fuel_modifier
 			if(fueluse > initial(fueluse)) //keep it at the max
 				fueluse = initial(fueluse)
-		return
+		return TRUE
 	else
 		if(on)
 			if(istype(W, /obj/item/natural/dirtclod))
